requirements.txt
requests~=2.26.0
Flask~=2.0.1
waitress~=2.0.0
python-dateutil~=2.8.1
pydantic~=1.8.2
websocket-client~=1.2.1